moving part of the Extensions.kt file from the `presentation` module to the `features:clients:presentation` module
